% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a11y_panel.R
\name{a11y_panel}
\alias{a11y_panel}
\title{Accessibility panel}
\usage{
a11y_panel(
  dashboard_title,
  dashboard_link,
  date_prepared,
  date_reviewed,
  public_repo_link = NA,
  non_accessible_components = NA,
  specific_issues = NA
)
}
\arguments{
\item{dashboard_title}{Title of the host dashboard}

\item{dashboard_link}{URL for the host dashboard}

\item{date_prepared}{Date the statement was prepared in it's current form}

\item{date_reviewed}{Date the statement was last reviewed}

\item{public_repo_link}{URL for the dashboard repository}

\item{non_accessible_components}{String vector containing a list of non accessible components}

\item{specific_issues}{String vector containing descriptions of specific accessibility issues
that have been identified as part of testing}
}
\value{
shiny$tags$div element containing the html tags and content for the standard
accessibility statement
}
\description{
Accessibility panel
}
\examples{
a11y_panel(
"DfE Shiny template",
"26th November 2024",
"26th November 2024",
"https://department-for-education.shinyapps.io/dfe-shiny-template",
public_repo_link = "https://github.com/dfe-analytical-services/shiny-template",
)
}
