% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_format_project.R
\name{air_format_project}
\alias{air_format_project}
\title{Air Format Project}
\usage{
air_format_project()
}
\description{
formats the whole project using air
}
\examples{
\dontrun{
air_format_project()
}
}
