% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_formatter.R
\name{air_formatter}
\alias{air_formatter}
\title{Air - format files}
\usage{
air_formatter(target = ".", verbose = FALSE)
}
\arguments{
\item{target}{single file target for formatting}

\item{verbose}{Run in verbose mode}
}
\description{
formats the whole project or single file using air
}
\examples{
\dontrun{
air_formatter()
}
}
