% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R
\name{cookie_banner_server}
\alias{cookie_banner_server}
\title{cookie_banner_server}
\usage{
cookie_banner_server(
  id,
  input_cookies,
  parent_session,
  google_analytics_key = NULL,
  cookie_link_panel
)
}
\arguments{
\item{id}{Shiny tag shared with cookie_banner_ui()}

\item{input_cookies}{The cookie input passed from cookies.js (should always
be reactive(input$cookies))
Should always be set to reactive(input$cookie_consent_clear).}

\item{parent_session}{This should be the R Shiny app session}

\item{google_analytics_key}{Provide the GA 10 digit key of the form
"ABCDE12345"}

\item{cookie_link_panel}{name of the navlistPanel that the cookie banner
provides a link to, usually "cookies_panel_ui"}
}
\description{
cookie_banner_server() provides the server module to be used alongside
cookie_banner_ui(). Place cookie_banner_server() as a call in your server.R
file to provide the server functions to control users being able to accept or
reject cookie consent for the provision of Google Analytics tracking on DfE
R-Shiny dashboards.
}
\examples{
\dontrun{
output$cookie_status <- dfeshiny::cookie_banner_server(
  "cookies",
  input_cookies = reactive(input$cookies),
  parent_session = session,
  google_analytics_key = "ABCDE12345",
  cookie_link_panel = "cookies_panel_ui"
)
}
}
