% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R
\name{cookie_banner_ui}
\alias{cookie_banner_ui}
\title{cookie_banner_ui}
\usage{
cookie_banner_ui(id, name = "DfE R-Shiny dashboard template")
}
\arguments{
\item{id}{Shiny tag shared with cookie_banner_server()}

\item{name}{Name of the dashboard on which the cookie authorisation is being
applied}
}
\value{
shiny::tags$div()
}
\description{
This function provides a cookie authorisation banner on DfE R-Shiny
dashboards for users to be able to accept or reject cookies. The server side
functionality is provided by cookie_banner_server(), whilst users will also
need to include the dfe_cookie_script() function in their ui.R file.
}
\examples{
cookie_banner_ui("cookies", name = "My DfE R-Shiny data dashboard")
}
