% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies_panel.R
\name{cookies_panel}
\alias{cookies_panel}
\title{Cookies panel}
\usage{
cookies_panel(cookie_status_output = "cookie_status")
}
\arguments{
\item{cookie_status_output}{Name of cookie status output object, often
"cookie_status"}
}
\value{
a standardised panel for a public R Shiny dashboard in DfE
}
\description{
Create the standard DfE R-Shiny cookies dashboard panel.
}
\examples{
cookies_panel(
  cookie_status_output = "cookie_status"
)
}
