% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies_panel.R
\name{cookies_panel_server}
\alias{cookies_panel_server}
\title{cookies_panel_server}
\usage{
cookies_panel_server(id)
}
\arguments{
\item{id}{ID shared with cookies_panel_ui()}
}
\description{
Create the server module of DfE R-Shiny cookies dashboard panel to be used
alongside cookies_panel_ui().
}
\examples{
\dontrun{
cookies_panel_server(
  "cookies_panel"
)
}
}
