% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dashboard.R
\name{create_dashboard}
\alias{create_dashboard}
\title{Initialize the DFE Shiny Template Setup}
\usage{
create_dashboard(
  path = "./",
  google_analytics_key = "XXXXXXXXX",
  dashboard_url = "https://department-for-education.shinyapps.io/dashboard-name/",
  site_title = "Dashboard name",
  parent_pub_name = "Publication name",
  parent_publication =
    "https://explore-education-statistics.service.gov.uk/find-statistics/publication-name",
  team_email = "team.name@education.gov.uk",
  repo_name = "https://github.com/dfe-analytical-services/dashboard-name",
  feedback_form_url = ""
)
}
\arguments{
\item{feedback_form_url}{}
}
\value{
No return value. The function initializes the complete template.
}
\description{
This function sets up a complete DFE Shiny template by:
\enumerate{
\item Cloning the Shiny template repository.
\item Installing a Git pre-commit hook.
\item Setting up a GitHub Actions workflow for deployment.
}
}
\details{
This function runs \code{init_app()}, \code{init_commit_hooks()}, and
\code{init_workflow()}
in sequence to set up a working environment for a Shiny application.
}
\examples{
\dontrun{
init_template()
init_template(dashboard_name = "my-dashboard")
}
}
