% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{custom_disconnect_message}
\alias{custom_disconnect_message}
\title{Custom disconnect message}
\usage{
custom_disconnect_message(
  refresh = "Refresh page",
  links = NULL,
  publication_name = NULL,
  publication_link = NULL
)
}
\arguments{
\item{links}{A list of mirros or alternative links to the dashboard}

\item{publication_name}{The parent publication name}

\item{publication_link}{The link to the publication on Explore Education Statistics}
}
\description{
Create the html overlay panel to appear when RSConnect disconnects
}
