% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{customDisconnectMessage}
\alias{customDisconnectMessage}
\title{Custom disconnect message}
\usage{
customDisconnectMessage(
  refresh = "Refresh page",
  links = sites_list,
  publication_name = ees_pub_name,
  publication_link = ees_publication
)
}
\arguments{
\item{links}{A list of mirros or alternative links to the dashboard}

\item{publication_name}{The parent publication name}

\item{publication_link}{The link to the publication on Explore Education Statistics}
}
\description{
Create the html overlay panel to appear when RSConnect disconnects
}
