% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps_shintest2.R
\name{deps_shinytest2}
\alias{deps_shinytest2}
\title{deps_shinytest2}
\usage{
deps_shinytest2()
}
\description{
This function installs the dependencies required by shinytest2.
}
\examples{
if (interactive()) {
  deps_shinytest2()
}
}
