% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe_content_links.R
\name{dfe_content_links}
\alias{dfe_content_links}
\title{Content links}
\usage{
dfe_content_links(links_list)
}
\arguments{
\item{links_list}{List of links for the navigation panel}
}
\description{
Content links for the right hand navigation list panel
}
\examples{
dfe_content_links(c("Data", "Information"))
}
