% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe_reactable.R
\name{dfe_reactable}
\alias{dfe_reactable}
\title{Department for Education Reactable Wrapper}
\usage{
dfe_reactable(data, ...)
}
\arguments{
\item{data}{A data frame to display in the table.}

\item{...}{Arguments passed to \code{shinyGovstyle::govReactable()}.}
}
\value{
The result of \code{shinyGovstyle::govReactable(...)}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

A wrapper around \code{shinyGovstyle::govReactable()}, preserving the function
here in dfeshiny for backwards compatibility in upcoming versions only.

All logic and documentation is now in \code{shinyGovstyle::govReactable()}, so
please refer to that function for more information, and update your code
to use that function directly.
}
\seealso{
\code{\link[shinyGovstyle:govReactable]{shinyGovstyle::govReactable()}}
}
