% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_link.R
\name{external_link}
\alias{external_link}
\title{External link}
\usage{
external_link(...)
}
\arguments{
\item{...}{Arguments passed to \code{shinyGovstyle::external_link()}.}
}
\value{
The result of \code{shinyGovstyle::external_link(...)}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

A wrapper around \code{shinyGovstyle::external_link()}, preserving the function
here in dfeshiny for backwards compatibility in upcoming versions only.

All logic and documentation is now in \code{shinyGovstyle::external_link()}, so
please refer to that function for more information, and update your code
to use that function directly.
}
\seealso{
\code{\link[shinyGovstyle:external_link]{shinyGovstyle::external_link()}}
}
