% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footer_dfe_auto.R
\name{footer_dfe}
\alias{footer_dfe}
\title{Footer Function}
\usage{
footer_dfe(full = FALSE)
}
\arguments{
\item{full}{Whenever you want to have blank footer or official gov version.
Defaults to \code{FALSE}}
}
\value{
a footer html shiny object
}
\description{
This function create a gov style footer for your page
}
\examples{
if (interactive()) {
  ui <- fluidPage(
    shinyGovstyle::header(
      main_text = "Example",
      secondary_text = "User Examples",
      logo = "shinyGovstyle/images/moj_logo.png"
    ),
    shinyGovstyle::banner(
      inputId = "banner", type = "beta", "This is a new service"
    ),
    tags$br(),
    tags$br(),
    footer_dfe(full = TRUE)
  )

  server <- function(input, output, session) {
    observeEvent(input$link_to_accessibility, {
      updateTabsetPanel(session, "navlistPanel", selected = "Accessibility")
    })
  }

  shinyApp(ui = ui, server = server)
}
}
\keyword{footer}
