% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{header}
\alias{header}
\title{DfE header banner}
\usage{
header(header, ...)
}
\arguments{
\item{header}{Text to use for the header of the dashboard}

\item{...}{Further arguments passed to shinyGovstyle::header()}
}
\value{
a header html shiny object
}
\description{
This function uses \code{shinyGovstyle::header()} to create a header banner
using the DfE logo.
}
\examples{

if (interactive()) {
  ui <- fluidPage(
    dfeshiny::header(
      header = "User Examples"
    )
  )

  server <- function(input, output, session) {}

  shinyApp(ui = ui, server = server)
}
}
\seealso{
\code{\link[shinyGovstyle:header]{shinyGovstyle::header()}}
}
