% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_paragraph_tags.R
\name{html_paragraph_tags}
\alias{html_paragraph_tags}
\title{Create \code{shiny::tagList()} for HTML paragraph}
\usage{
html_paragraph_tags(heading, body)
}
\arguments{
\item{heading}{A single vector or a combined vector wrapped in
\code{shiny::tagList()} instead of c() for the heading of your paragraph}

\item{body}{A single vector or a combined vector wrapped in
\code{shiny::tagList()} instead of c() for the body of your paragraph}
}
\value{
A list of HTML tags that contain and h2 heading and a paragraph body.
}
\description{
Create a list of tags that generate a HTML paragraph.
This generates a header using shiny::tags$h2()
and a paragraph body using shiny::tags$p()

This is to be used with \code{dfeshiny::suport_panel()} to add extra sections to a public R Shiny
dashboard in DfE
}
\examples{

# Example for text heading and body
html_paragraph_tags(
  heading = "Heading test",
  body = "This is a body text test"
)

# Example for text heading and text with other elements in the body


html_paragraph_tags(
  heading = "Heading test",
  body = shiny::tagList(
    "This is a body text test. Please contact us at",
    dfeshiny::external_link(
      href = paste0("mailto:", "team@education.gov.uk"),
      link_text = "team@education.gov.uk",
      add_warning = FALSE
    )
  )
)

}
\seealso{
\code{\link[=support_panel]{support_panel()}}
}
