% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{init_analytics}
\alias{init_analytics}
\title{init_analytics}
\usage{
init_analytics(ga_code, create_file = TRUE)
}
\arguments{
\item{ga_code}{The Google Analytics code for the dashboard}

\item{create_file}{Boolean TRUE or FALSE, default is TRUE, false will return
the HTML in the console and is used mainly for testing or comparisons}
}
\value{
TRUE if written, FALSE if not, character vector of HTML if create_file = FALSE
}
\description{
Creates the google-analytics.html script in order to allow the activation of
analytics via GA4. For the full steps required to set up analytics, please
refer to the documentation in the README.
}
\examples{
init_analytics(ga_code = "0123456789")
}
