% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_app.R
\name{init_app}
\alias{init_app}
\title{Load DFE Template R Shiny App into the Working Directory}
\usage{
init_app()
}
\value{
No return value. The function downloads the Shiny app template
into the current directory.
}
\description{
This function clones a template R Shiny application from the repository
"https://github.com.mcas.ms/dfe-analytical-services/shiny-template"
into the current working directory.
}
\details{
The function checks if Git is installed before attempting
to clone the repository. If Git is not found, it stops with an error message.
The downloaded repository contains a pre-configured R Shiny app template,
which can be used as a starting point for development.
}
\examples{
\dontrun{
init_app()
}
}
