% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dashboard.R
\name{init_base_app}
\alias{init_base_app}
\title{Initialize a Shiny App Project Structure}
\usage{
init_base_app(
  path = "./",
  google_analytics_key = "XXXXXXXXX",
  dashboard_url = "https://department-for-education.shinyapps.io/dashboard-name/",
  site_title = "Dashboard name",
  parent_pub_name = "Publication name",
  parent_publication =
    "https://explore-education-statistics.service.gov.uk/find-statistics/publication-name",
  team_email = "team.name@education.gov.uk",
  repo_name = "https://github.com/dfe-analytical-services/dashboard-name",
  feedback_form_url = ""
)
}
\arguments{
\item{path}{A character string specifying the root directory
where the app structure should be created. Defaults to the current
working directory (\code{"."}).}

\item{feedback_form_url}{}
}
\value{
No return value. The function is called for
its side effects (i.e., file and folder creation).
}
\description{
Creates the basic file and directory structure for a Shiny application,
including:\code{global.R}, \code{ui.R}, \code{server.R}, a \verb{data/} folder,
and a \verb{tests/testthat/} folder.
}
\examples{
\dontrun{
init_app()              # Creates structure in current directory
init_app("myShinyApp")  # Creates structure in 'myShinyApp' directory
}

}
\keyword{internal}
