% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dashboard.R
\name{init_global}
\alias{init_global}
\title{Initialise global.R}
\usage{
init_global(
  path = "./",
  google_analytics_key = "XXXXXXXXX",
  dashboard_url = "https://department-for-education.shinyapps.io/dashboard-name/",
  site_title = "Dashboard name",
  parent_pub_name = "Publication name",
  parent_publication =
    "https://explore-education-statistics.service.gov.uk/find-statistics/publication-name",
  team_email = "team.name@education.gov.uk",
  repo_name = "https://github.com/dfe-analytical-services/dashboard-name",
  feedback_form_url = ""
)
}
\arguments{
\item{feedback_form_url}{}
}
\description{
Initialise global.R
}
\keyword{internal}
