% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_commit.R
\name{init_hooks}
\alias{init_hooks}
\title{Initialize Pre-commit Hook}
\usage{
init_hooks()
}
\value{
The path to the pre-commit hook file if created successfully.
}
\description{
Creates a Git pre-commit hook that runs the \code{commit_hooks} function.
}
