% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dashboard.R
\name{init_server}
\alias{init_server}
\title{Initialise server.R}
\usage{
init_server(path = ".", verbose = FALSE)
}
\arguments{
\item{path}{The root directory where the app should be created. String, default: \code{"."}.}

\item{verbose}{RUn in verbose mode}
}
\description{
Initialise server.R
}
\examples{
\dontrun{
init_server()
}
}
\keyword{internal}
