% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_app.R
\name{init_workflow}
\alias{init_workflow}
\title{Initialize GitHub Actions Workflow for Shiny App Deployment}
\usage{
init_workflow(dashboard_name)
}
\arguments{
\item{dashboard_name}{A character string specifying the dashboard name.
default value is \code{"dfe-shiny-template"}.}
}
\value{
No return value. The function sets up the workflow and configuration.
}
\description{
This function sets up a GitHub Actions workflow for deploying a Shiny app.
It ensures that the \verb{.github/workflows/} directory exists, downloads a
deployment workflow file, and creates a configuration YAML file.
}
\details{
If the \verb{.github/workflows/} directory does not exist, it is created.
The function then downloads the deployment workflow file and places it inside
\verb{.github/workflows/}. Additionally, it generates a YAML file containing
deployment settings, allowing customization of the \code{dashboard_name}.
}
\examples{
\dontrun{
init_workflow()
init_workflow(dashboard_name = "my-custom-dashboard")
}
}
