% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisers.R
\name{initialise_analytics}
\alias{initialise_analytics}
\title{initialise_analytics}
\usage{
initialise_analytics(ga_code)
}
\arguments{
\item{ga_code}{The Google Analytics code}
}
\value{
TRUE if written, FALSE if not
}
\description{
initialise_analytics
}
\examples{
initialise_analytics(ga_code = "0123456789")
}
