% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_commit.R
\name{pre_commit_innit}
\alias{pre_commit_innit}
\title{Initialize Pre-commit Hook}
\usage{
pre_commit_innit(skip_analytics_key_check = FALSE)
}
\arguments{
\item{skip_analytics_key_check}{Logical. If TRUE, skips \code{check_analytics_key} in the hook.}
}
\value{
The path to the pre-commit hook file if created successfully.
}
\description{
Creates a Git pre-commit hook that runs the \code{commit_hooks} function.
If \code{skip_analytics_key_check} is set to TRUE, it will pass that argument to the script.
}
