% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_tags.R
\name{section_tags}
\alias{section_tags}
\title{Create \code{shiny::tagList()} for HTML section}
\usage{
section_tags(heading = NULL, body, h_level = "h2")
}
\arguments{
\item{heading}{Optional - A single text string or a \code{shiny::tagList()} object
for the heading of your paragraph}

\item{body}{A single text string or a \code{shiny::tagList()} object
for the body of your paragraph}

\item{h_level}{Specify the level for your heading, if you choose to include
one.Default is "h2". Available options are "h2","h3" or "h4".}
}
\value{
A list of HTML tags that contain and h2 heading and a paragraph body.
}
\description{
Create a list of tags that generate a HTML section.
This generates a header using \code{shiny::tags$h2()}
and a paragraph body using \code{shiny::tags$p()}

This is to be used with functions in the \code{dfeshiny} package
to add extra sections to a public R Shiny dashboard in DfE
}
\examples{

# Example for text heading and body
section_tags(
  heading = "Heading test",
  body = "This is a body text test"
)

# Example for using a different heading level
section_tags(
  heading = "Heading test",
  body = "This is a body text test",
  h_level = "h3"
)


# Example for text heading and text with other elements in the body


section_tags(
  heading = "Heading test",
  body = shiny::tagList(
    "This is a body text test. Please contact us at",
    dfeshiny::external_link(
      href = paste0("mailto:", "team@education.gov.uk"),
      link_text = "team@education.gov.uk",
      add_warning = FALSE
    )
  )
)

}
\seealso{
\code{\link[=support_panel]{support_panel()}}

\link[htmltools:builder]{htmltools::tags}

\code{\link[htmltools:tagList]{htmltools::tagList()}}
}
