% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmark.R
\name{set_bookmark_include}
\alias{set_bookmark_include}
\title{DfE bookmark include}
\usage{
set_bookmark_include(session, input, bookmarking_whitelist)
}
\arguments{
\item{session}{default shiny session}

\item{input}{should be input, it pulls in the shiny inputs}

\item{bookmarking_whitelist}{list of inputs to include in bookmark}
}
\value{
a bookmark with relevant inputs included
}
\description{
This function allows for a whitelist of included inputs
for bookmarking
}
\examples{
# You will need a line such as this in your global.R script ================
bookmarking_whitelist <- c("navlistPanel", "tabsetpanels")

# In the server.R script ===================================================
shiny::observe({
  set_bookmark_include(session, input, bookmarking_whitelist)
})

shiny::observe({
  # Trigger this observer every time an input changes
  reactiveValuesToList(input)
  session$doBookmark()
})
}
