% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromote_setup.R
\name{shinytest2_setup}
\alias{shinytest2_setup}
\title{shinytest2 set up}
\usage{
shinytest2_setup()
}
\description{
This sets up chromote for use with shinytest2. We've had some issues with
changes to Chrome headless mode and other issues with connecting to Chrome /
Edge, so this aims to sidestep having to provide convoluted guidance to
analysts and allow them to run a single function to get set up.
}
\examples{
shinytest2_setup()
}
