% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_commit.R
\name{style_code}
\alias{style_code}
\title{Check and Apply Code Styling}
\usage{
style_code()
}
\value{
TRUE if changes were made, FALSE otherwise.
}
\description{
Styles R code in the \code{R} directory and detects changes.
If changes are made, warns the user before committing.
}
