% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_yml.R
\name{use_yml_deploy}
\alias{use_yml_deploy}
\title{use_yml_deploy}
\usage{
use_yml_deploy(app_name = NULL, mirror = FALSE, dev = FALSE)
}
\arguments{
\item{app_name}{Name of the app to use as the url slug}

\item{mirror}{If TRUE, create a mirror site}

\item{dev}{If TRUE, create a development site from the development branch}
}
\description{
use_yml_deploy
}
