% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_date}
\alias{validate_date}
\title{Date validation}
\usage{
validate_date(date)
}
\arguments{
\item{date}{Element to be validated as a date}
}
\value{
String containing validation message
}
\description{
Date validation
}
