% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a11y_panel.R
\name{a11y_panel}
\alias{a11y_panel}
\title{Accessibility panel}
\usage{
a11y_panel(
  dashboard_title,
  dashboard_url,
  date_tested,
  date_prepared,
  date_reviewed,
  date_template_reviewed = "12 March 2024",
  repo_url = NA,
  non_accessible_components =
    c("Keyboard navigation through the interactive charts is currently limited",
    "Alternative text in interactive charts is limited to titles"),
  specific_issues =
    c("Charts have non-accessible components that are inaccessible for keyboard users.",
    "Chart tooltips are not compativle with screen reader use.",
    "Some decorative images are not labelled appropriately as yet.",
    "Some links are not appropriately labelled.")
)
}
\arguments{
\item{dashboard_title}{Title of the host dashboard}

\item{dashboard_url}{URL for the host dashboard}

\item{date_tested}{Date the application was last tested}

\item{date_prepared}{Date the statement was prepared in it's current form}

\item{date_reviewed}{Date the statement was last reviewed}

\item{date_template_reviewed}{Date the underlying template was reviewed
(default: 12th March 2024)}

\item{repo_url}{URL for the dashboard repository}

\item{non_accessible_components}{String vector containing a list of non accessible components}

\item{specific_issues}{String vector containing descriptions of specific accessibility issues
that have been identified as part of testing}
}
\value{
shiny$tags$div element containing the html tags and content for the standard
accessibility statement
}
\description{
Accessibility panel
}
\examples{
a11y_panel(
  "DfE Shiny template",
  "https://department-for-education.shinyapps.io/dfe-shiny-template",
  "25th April 2024",
  "26th April 2024",
  "2nd November 2024",
  repo_url = "https://github.com/dfe-analytical-services/shiny-template",
)
}
