% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_format_project.R
\name{air_format_project}
\alias{air_format_project}
\title{Air Format Project}
\usage{
air_format_project(target = ".")
}
\arguments{
\item{target}{single file target for formatting}
}
\description{
formats the whole project or single file using air
}
\examples{
\dontrun{
air_format_project()
}
}
