% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_commit.R
\name{commit_hooks}
\alias{commit_hooks}
\title{Commit Hooks for Data Validation, Analytics Check, and Code Styling}
\usage{
commit_hooks(skip_analytics_key_check = FALSE)
}
\arguments{
\item{skip_analytics_key_check}{Logical. If TRUE, skips \code{check_analytics_key}.}
}
\value{
TRUE if all checks pass, FALSE otherwise.
}
\description{
This wrapper function executes a series of pre-commit checks, including:
\itemize{
\item Data file validation (\code{data_checker})
\item Google Analytics key validation (\code{check_analytics_key}, can be skipped)
\item Code styling check (\code{style_code})
}
}
