% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R
\name{cookie_banner_server}
\alias{cookie_banner_server}
\title{cookie_banner_server}
\usage{
cookie_banner_server(
  id = "cookies_banner",
  input_cookies = reactive(input$cookies),
  parent_session = session,
  google_analytics_key = NULL,
  cookie_link_panel = "cookies_panel_ui"
)
}
\arguments{
\item{id}{Shiny tag shared with cookie_banner_ui(), can be any string set by
the user as long as it matches the id in the cookie_banner_ui()}

\item{input_cookies}{The cookie input passed from cookies.js (should always
be \code{reactive(input$cookies)})}

\item{parent_session}{This should be the R Shiny app session}

\item{google_analytics_key}{Provide the GA 10 digit key of the form
"ABCDE12345"}

\item{cookie_link_panel}{name of the navlistPanel that the cookie banner
provides a link to, usually "cookies_panel_ui"}
}
\description{
cookie_banner_server() provides the server module to be used alongside
cookie_banner_ui(). Place cookie_banner_server() as a call in your server.R
file to provide the server functions to control users being able to accept or
reject cookie consent for the provision of Google Analytics tracking on DfE
R-Shiny dashboards.
}
\examples{
if (interactive()) {
  # This example shows how to use the full family of cookie functions together
  # This will be in your global.R script =====================================

  library(shiny)
  library(shinyjs)
  library(dfeshiny)
  ga_key <- "ABCDE12345"

  # This will be what is in your ui.R script =================================

  ui <- fluidPage(
    # Place these lines above your header ------------------------------------
    dfe_cookie_script(),
    useShinyjs(),
    cookie_banner_ui(name = "My DfE R-Shiny data dashboard"),

    # Place the cookies panel under the header but in the main content -------
    cookies_panel_ui(google_analytics_key = ga_key)
  )

  # This will be in your server.R file =======================================

  server <- function(input, output, session) {
    # Server logic for the pop up banner, can be placed anywhere in server.R -
    output$cookie_status <- dfeshiny::cookie_banner_server(
      input_cookies = reactive(input$cookies),
      google_analytics_key = ga_key
    )

    # Server logic for the panel, can be placed anywhere in server.R ---------
    cookies_panel_server(
      input_cookies = reactive(input$cookies),
      google_analytics_key = ga_key
    )
  }

  # How to run the minimal app given in this example =========================
  shinyApp(ui, server)
}
}
\seealso{
Other cookies: 
\code{\link{cookie_banner_ui}()},
\code{\link{cookies}}
}
\concept{cookies}
