% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R
\name{cookie_banner_ui}
\alias{cookie_banner_ui}
\title{cookie_banner_ui}
\usage{
cookie_banner_ui(id, name = "DfE R-Shiny dashboard template")
}
\arguments{
\item{id}{Shiny tag shared with cookie_banner_server()}

\item{name}{Name of the dashboard on which the cookie authorisation is being
applied}
}
\value{
shiny::tags$div()
}
\description{
This function provides a cookie authorisation banner on DfE R-Shiny
dashboards for users to be able to accept or reject cookies. The server side
functionality is provided by cookie_banner_server(), whilst users will also
need to include the dfe_cookie_script() function in their ui.R file.

To use the banner, first copy the file cookie-consent.js from the js/ folder
of this package to the www/ folder of your R-Shiny app. Then add the
following lines to your ui.R:
dfe_cookie_script()
cookie_banner_ui("cookies", name = "My DfE R-Shiny data dashboard")

And add the following in server.R (after updating the google analytics key):
output$cookie_status <- dfeshiny::cookie_banner_server(
"cookies",
input.cookies = reactive(input$cookies),
input.remove = reactive(input$remove),
parent_session = session,
google_analytics_key = "ABCDE12345"
)

Note that you should also include dfeshiny::support_panel() in your
navListPanel()
}
\examples{
cookie_banner_ui("cookies", name = "My DfE R-Shiny data dashboard")
}
