% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R
\name{cookieBannerServer}
\alias{cookieBannerServer}
\title{cookieBannerServer}
\usage{
cookieBannerServer(id, input.cookies = NULL, input.remove = NULL)
}
\arguments{
\item{input.remove}{}
}
\description{
cookieBannerServer
}
