% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R
\name{cookieBannerServer}
\alias{cookieBannerServer}
\title{cookieBannerServer}
\usage{
cookieBannerServer(id, input.cookies, input.remove, parent_session)
}
\arguments{
\item{id}{Shiny tag shared with cookieBannerUI()}

\item{input.cookies}{The cookie input passed from cookies.js (should always
be reactive(input$cookies))}

\item{input.remove}{The state of the cookie reset button provided by
dfeshiny::support_panel(). Should always be set to reactive(input$remove).}
}
\description{
cookieBannerServer() provides the server module to be used alongside
cookieBannerUI(). Place cookieBannerServer() as a call in your server.R file
to provide the server functions to control users being able to accept or
reject cookie consent for the provision of Google Analytics tracking on DfE
R-Shiny dashboards.

To use the banner, first copy the file cookie-consent.js from the js/ folder
of this package to the www/ folder of your R-Shiny app. Then add the
following lines to your ui.R:
dfe_cookie_script()
cookieBannerUI("cookies", name = "My DfE R-Shiny data dashboard")

And add the following in server.R:
output$cookie_status <- dfeshiny::cookieBannerServer(
"cookies",
input.cookies = reactive(input$cookies),
input.remove = reactive(input$remove),
parent_session = session
)

Note that you should also include dfeshiny::support_panel() in your
navListPanel()
}
\examples{
output$cookie_status <- cookieBannerServer(
  "cookies",
  input.cookies = reactive(input$cookies),
  input.remove = reactive(input$remove),
  parent_session = session
)
}
