% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R
\name{cookies}
\alias{cookies}
\alias{dfe_cookies_script}
\title{dfe_cookies_script}
\usage{
dfe_cookies_script()
}
\value{
shiny::tags$head()
}
\description{
Calls in JavaScript dependencies to the shiny app used to set and unset the
cookies. Function should be placed in the ui.R script.
}
\examples{
if (interactive()) {
  # This example shows how to use the full family of cookies functions together
  # This assumes you are using bslib for your UI (recommended)
  # This will be in your global.R script ===================================

  library(shiny)
  library(shinyjs)
  library(bslib)
  library(dfeshiny)
  google_analytics_key <- "ABCDE12345"

  # This will be what is in your ui.R script ===============================

  ui <- bslib::page_fluid(
    # Place these lines above your header ----------------------------------
    useShinyjs(),
    dfe_cookies_script(),
    cookies_banner_ui(name = "My DfE R Shiny data dashboard"),

    # Place the cookies panel under the header but in the main content -----
    bslib::navset_tab(
      id = "left_nav",

      # UI panels ==========================================================
      bslib::nav_panel(
        title = "My amazing dashboard",
        id = "so_great",
        shiny::tags$h1("Welcome to my amazing dashboard")
      ),
      bslib::nav_panel(
        title = "Cookies",
        id = "cookies_panel",
        dfeshiny::cookies_panel_ui(google_analytics_key = ga_key)
      )
    )
  )

  # This will be in your server.R file =====================================

  server <- function(input, output, session) {
    output$cookies_status <- dfeshiny::cookies_banner_server(
      input_cookies = reactive(input$cookies),
      google_analytics_key = google_analytics_key,
      parent_session = session
    )

    cookies_panel_server(
      input_cookies = reactive(input$cookies),
      google_analytics_key = google_analytics_key
    )

    shiny::observeEvent(input$`cookies_banner-cookies_link`, {
      shiny::updateTabsetPanel(session, "left_nav", selected = "cookies_panel")
    })
  }

  # How to run the minimal app given in this example =======================
  shinyApp(ui, server)

  # ========================================================================
  # Example 2: Plain Shiny UI
  # This will be in your global.R script ===================================

  library(shiny)
  library(shinyjs)
  library(dfeshiny)
  google_analytics_key <- "ABCDE12345"

  # This will be what is in your ui.R script ===============================

  ui_plain <- fluidPage(
    # Place these lines above your header ----------------------------------
    useShinyjs(),
    dfe_cookies_script(),
    cookies_banner_ui(name = "My DfE R-Shiny data dashboard"),

    # Place the cookies panel under the header but in the main content -----
    # Example of placing as a panel within navlistPanel
    shiny::navlistPanel(
      "",
      id = "navlistPanel",
      widths = c(2, 8),
      well = FALSE,
      ## Cookies panel -----------------------------------------------------
      shiny::tabPanel(
        value = "cookies_panel_ui",
        "Cookies",
        cookies_panel_ui(google_analytics_key = google_analytics_key)
      )
    )
  )

  # This will be in your server.R file =====================================

  server_plain <- function(input, output, session) {
    # Server logic for the pop up banner, can be placed anywhere in server.R -
    output$cookies_status <- dfeshiny::cookies_banner_server(
      input_cookies = reactive(input$cookies),
      google_analytics_key = google_analytics_key,
      parent_session = session
    )

    # Server logic for the panel, can be placed anywhere in server.R -------
    cookies_panel_server(
      input_cookies = reactive(input$cookies),
      google_analytics_key = google_analytics_key
    )
  }

  # How to run the minimal app given in this example =======================
  shinyApp(ui_plain, server_plain)
}
}
\seealso{
Other cookies: 
\code{\link{cookies_banner_server}()},
\code{\link{cookies_banner_ui}()},
\code{\link{cookies_panel_server}()},
\code{\link{cookies_panel_ui}()}
}
\concept{cookies}
