% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R
\name{cookies_panel_server}
\alias{cookies_panel_server}
\title{cookies_panel_server}
\usage{
cookies_panel_server(
  id = "cookies_panel",
  input_cookies = reactive(input$cookies),
  google_analytics_key = NULL
)
}
\arguments{
\item{id}{ID shared with cookies_panel_ui()}

\item{input_cookies}{The cookie input passed from cookies.js (should always
be reactive(input$cookies))}

\item{google_analytics_key}{Provide the GA 10 digit key of the form
"ABCDE12345"}
}
\description{
Create the server module of DfE R-Shiny cookies dashboard panel to be used
alongside cookies_panel_ui().
}
\examples{
# This example shows how to use the full family of cookie functions together
# This will be in your global.R script =====================================

library(shiny)
library(shinyjs)
library(dfeshiny)
ga_key <- "ABCDE12345"

# This will be what is in your ui.R script =================================

ui <- fluidPage(
  # Place these lines above your header ------------------------------------
  dfe_cookie_script(),
  useShinyjs(),
  cookie_banner_ui(name = "My DfE R-Shiny data dashboard"),

  # Place the cookies panel under the header but in the main content -------
  cookies_panel_ui(google_analytics_key = ga_key)
)

# This will be in your server.R file =======================================

server <- function(input, output, session) {
  # Server logic for the pop up banner, can be placed anywhere in server.R -
  output$cookie_status <- dfeshiny::cookie_banner_server(google_analytics_key = ga_key)

  # Server logic for the panel, can be placed anywhere in server.R ---------
  cookies_panel_server(google_analytics_key = ga_key)
}

# How to run the minimal app give in this example ==========================
if (interactive()) {
  shinyApp(ui, server)
}

}
