% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies_panel.R
\name{cookies_panel_ui}
\alias{cookies_panel_ui}
\title{cookies_panel_ui}
\usage{
cookies_panel_ui(
  id,
  cookie_status_output = "cookie_status",
  google_analytics_key = NULL
)
}
\arguments{
\item{id}{ID shared with cookies_panel_server()}

\item{cookie_status_output}{Name of cookie status output object, often
"cookie_status"}

\item{google_analytics_key}{Provide the GA 10 digit key of the form
"ABCDE12345"}
}
\value{
a standardised panel for a public R Shiny dashboard in DfE
}
\description{
Create the standard DfE R-Shiny cookies dashboard panel in the ui. The server
side functionality is provided by cookies_panel_server()
}
\examples{
\dontrun{
cookies_panel_ui(
  id = "cookies_panel",
  cookie_status_output = "cookie_status",
  google_analytics_key = "ABCDE12345"
)
}
}
