% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dashboard.R
\name{create_dashboard}
\alias{create_dashboard}
\title{Initialize the DFE Shiny Template Setup}
\usage{
create_dashboard(
  path = "./",
  site_title = "dfeshiny template",
  google_analytics_key = "XXXXXXXXXX",
  dashboard_url = "https://department-for-education.shinyapps.io/dfeshiny-template/",
  parent_pub_name = "Publication name",
  parent_publication =
    paste0("https://explore-education-statistics.service.gov.uk/find-statistics/",
    "publication-name"),
  team_email = "explore.statistics@education.gov.uk",
  repo_name = "https://github.com/dfe-analytical-services/dfeshiny",
  feedback_form_url = "",
  verbose = FALSE
)
}
\arguments{
\item{path}{The root directory where the app should be created. String, default: \code{"."}.}

\item{site_title}{Site title for the dashboard. String, default: \code{"Shiny template"}}

\item{google_analytics_key}{10 digit Google Analytics key if available. String, default:
\code{"XXXXXXXXXX"}}

\item{team_email}{Email address for support contact. String, default:
\code{"explore.statistics@education.gov.uk"}}

\item{repo_name}{The repository URL, must be a valid URL for the dfe-analytical-services GitHub
area or the dfe-gov-uk Azure DevOps. String, default:
\code{"https://github.com/dfe-analytical-services/dfeshiny"}}

\item{verbose}{}
}
\value{
No return value. The function initializes the complete template.
}
\description{
This function sets up a complete DFE Shiny template by:
\enumerate{
\item Cloning the Shiny template repository.
\item Installing a Git pre-commit hook.
\item Setting up a GitHub Actions workflow for deployment.
}
}
\details{
This function runs \code{init_app()}, \code{init_commit_hooks()}, and
\code{init_workflow()}
in sequence to set up a working environment for a Shiny application.
}
\examples{
\dontrun{
create_dashboard()
}
}
