% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_disconnect_message.R
\name{custom_disconnect_message}
\alias{custom_disconnect_message}
\title{Custom disconnect message}
\usage{
custom_disconnect_message(
  refresh = "Refresh page",
  dashboard_title = NULL,
  links = NULL,
  publication_name = NULL,
  publication_link = NULL
)
}
\arguments{
\item{refresh}{the text to appear that will refresh the page when clicked}

\item{dashboard_title}{Title of the dashboard}

\item{links}{A vector of possible URLs for the public site. Should mostly just be a single URL,
but can be two URLs if an overflow site has been set up}

\item{publication_name}{The parent publication name}

\item{publication_link}{The link to the publication on Explore Education
Statistics}
}
\value{
A HTML overlay panel that appears when a user loses connection to a DfE R Shiny
dashboard.
}
\description{
Create the HTML overlay panel to appear when a user loses connection to a dashboard.
}
\examples{
custom_disconnect_message(
  dashboard_title = "DfE Shiny Template",
  refresh = "Refresh page",
  links = c(
    "https://department-for-education.shinyapps.io/dfe-shiny-template",
    "https://department-for-education.shinyapps.io/dfe-shiny-template-overflow/"
  ),
  publication_name = "Explore Education Statistics Publication",
  publication_link =
    "https://explore-education-statistics.service.gov.uk/find-statistics/apprenticeships"
)

custom_disconnect_message(
  dashboard_title = "DfE Shiny Template",
  refresh = "Refresh page",
  links = c(
    "https://department-for-education.shinyapps.io/dfe-shiny-template"
  )
)

}
