% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_link.R
\name{external_link}
\alias{external_link}
\title{External link}
\usage{
external_link(href, link_text, add_warning = TRUE)
}
\arguments{
\item{href}{URL that you want the link to point to}

\item{link_text}{Text that will appear describing your link, must be
descriptive of the page you are linking to. Vague text like 'click here' or
'here' will cause an error.}

\item{add_warning}{Boolean for adding "(opens in new tab)" at the end of the
link text to warn users of the behaviour. Be careful and consider
accessibility before removing the visual warning.}
}
\value{
shiny.tag object
}
\description{
It is commonplace for external links to open in a new tab, and when we do
this we should be careful...

This function automatically adds the following to your link:
\itemize{
\item \code{target="_blank"} to open in new tab
\item \code{rel="noopener noreferrer"} to prevent reverse tabnabbing
}

By default this function also adds "(opens in new tab)" to your link text
to warn users of the behaviour.

This also adds "This link opens in a new tab" as a visually hidden span
element within the html outputted to warn non-visual users of the behaviour.

Related links (aware of the painful irony but couldn't make the
documentation work in any other way!)...

Government digital services guidelines on the use of links:
https://design-system.service.gov.uk/styles/links/

Anchor tag html element and its properties:
https://developer.mozilla.org/en-US/docs/Web/HTML/Element/a

Web Accessibility standards link text behaviour:
https://www.w3.org/TR/WCAG20-TECHS/G200.html

Reverse tabnabbing:
https://owasp.org/www-community/attacks/Reverse_Tabnabbing
}
\details{
Intentionally basic wrapper for html anchor elements making it easier to
create safe external links with standard and accessible behaviour. For more
information on how the tag is generated, see \code{\link[htmltools]{tags}}.
}
\examples{
external_link("https://shiny.posit.co/", "R Shiny")

external_link(
  "https://shiny.posit.co/",
  "R Shiny",
  add_warning = FALSE
)
}
