% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{header}
\alias{header}
\title{DfE header banner}
\usage{
header(secondary_text, secondary_link = "#")
}
\arguments{
\item{secondary_text}{Secondary header to supplement the main text}

\item{secondary_link}{Add a link for clicking on secondary header}
}
\value{
a header html shiny object
}
\description{
This function uses \code{shinyGovstyle::header()} to create a header banner
using the DfE logo.
}
\examples{

if (interactive()) {

  ui <- fluidPage(
    dfeshiny::header(
      secondary_text = "User Examples")
  )

  server <- function(input, output, session) {}

  shinyApp(ui = ui, server = server)
}
}
\seealso{
\code{\link[shinyGovstyle:header]{shinyGovstyle::header()}}
}
