% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{init_analytics}
\alias{init_analytics}
\title{init_analytics}
\usage{
init_analytics(ga_code)
}
\arguments{
\item{ga_code}{The Google Analytics code}
}
\value{
TRUE if written, FALSE if not
}
\description{
Creates the google-analytics.html script in order to allow the activation of
analytics via GA4. For the full steps required to set up analytics, please
refer to the documentation in the readme.
}
\examples{
init_analytics(ga_code = "0123456789")
}
