% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisers.R
\name{init_analytics}
\alias{init_analytics}
\title{init_analytics}
\usage{
init_analytics(ga_code)
}
\arguments{
\item{ga_code}{The Google Analytics code}
}
\value{
TRUE if written, FALSE if not
}
\description{
init_analytics
}
\examples{
init_analytics(ga_code = "0123456789")
}
