% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_app.R
\name{init_app}
\alias{init_app}
\title{Initialize a Shiny App Project Structure}
\usage{
init_app(path = ".")
}
\arguments{
\item{path}{A character string specifying the root directory
where the app structure should be created. Defaults to the current
working directory (\code{"."}).}
}
\value{
No return value. The function is called for
its side effects (i.e., file and folder creation).
}
\description{
Creates the basic file and directory structure for a Shiny application,
including:\code{global.R}, \code{ui.R}, \code{server.R}, a \verb{data/} folder,
and a \verb{tests/testthat/} folder.
}
\examples{
\dontrun{
init_app()              # Creates structure in current directory
init_app("myShinyApp")  # Creates structure in 'myShinyApp' directory
}

}
