% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dashboard.R
\name{init_base_app}
\alias{init_base_app}
\title{Initialize a Shiny App Project Structure}
\usage{
init_base_app(
  path = "./",
  google_analytics_key = "XXXXXXXXX",
  dashboard_url = "https://department-for-education.shinyapps.io/dashboard-name/",
  site_title = "Dashboard name",
  publication_name = "Publication name",
  publication_link =
    paste0("https://explore-education-statistics.service.gov.uk/find-statistics/",
    "publication-name"),
  team_email = "team.name@education.gov.uk",
  repo_name = "https://github.com/dfe-analytical-services/dashboard-name",
  form_url = "",
  verbose = FALSE
)
}
\arguments{
\item{path}{The root directory where the app should be created. String, default: \code{"."}.}

\item{google_analytics_key}{10 digit Google Analytics key if available. String, default:
\code{"XXXXXXXXXX"}}

\item{dashboard_url}{URL for the host dashboard}

\item{site_title}{Site title for the dashboard. String, default: \code{"dfeshiny template"}}

\item{publication_name}{The parent publication name}

\item{publication_link}{URL link to the parent publication}

\item{team_email}{Email address for support contact. String, default:
\code{"explore.statistics@education.gov.uk"}}

\item{repo_name}{The repository URL, must be a valid URL for the dfe-analytical-services GitHub
area or the dfe-gov-uk Azure DevOps. String, default:
\code{"https://github.com/dfe-analytical-services/dfeshiny"}}

\item{form_url}{URL for a feedback form for the dashboard}

\item{verbose}{RUn in verbose mode}
}
\value{
No return value. The function is called for
its side effects (i.e., file and folder creation).
}
\description{
Creates the basic file and directory structure for a Shiny application,
including:\code{global.R}, \code{ui.R}, \code{server.R}, a \verb{data/} folder,
and a \verb{tests/testthat/} folder.
}
\examples{
\dontrun{
init_base_app()              # Creates structure in current directory
}

}
\keyword{internal}
