% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dashboard.R
\name{init_commit_hooks}
\alias{init_commit_hooks}
\title{Initialize Git Pre-Commit Hook for a Shiny App}
\usage{
init_commit_hooks(path = "./")
}
\arguments{
\item{path}{The root directory where the app should be created. String, default: \code{"."}.}
}
\value{
No return value. The function installs the pre-commit hook.
}
\description{
This function checks if the current directory is a Git repository, then
downloads a pre-commit hook script from the
"dfe-analytical-services/shiny-template" repository and
places it in \code{.git/hooks/pre-commit}, overwriting if necessary.
}
\details{
The function verifies that the working directory is part of a
Git repository before proceeding. If Git is not installed or the directory
is not a Git repository, it stops with an error message.
The downloaded script ensures best practices before committing changes.
}
\examples{
\dontrun{
init_commit_hooks()
}
}
