% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dashboard.R
\name{init_ui}
\alias{init_ui}
\title{Initialise ui.R}
\usage{
init_ui(path = ".", team_email = "team.name@education.gov.uk")
}
\arguments{
\item{path}{The root directory where the app should be created. String, default: \code{"."}.}

\item{team_email}{Email address for support contact. String, default:
\code{"explore.statistics@education.gov.uk"}}
}
\description{
Initialise ui.R
}
\examples{
\dontrun{
init_ui()
}
}
\keyword{internal}
