% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_air.R
\name{install_air}
\alias{install_air}
\title{Air Install}
\usage{
install_air(update_global_settings = TRUE)
}
\arguments{
\item{update_global_settings}{auto update global settings, or don't}
}
\description{
checks for air installation status and installs it if
required, updating the global settings if selected
}
\examples{
\dontrun{
install_air()
}
}
