% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_panels.R
\name{support_panel}
\alias{support_panel}
\title{Support panel}
\usage{
support_panel(
  team_email = "",
  repo_name = "",
  publication_name = "",
  publication_stub = "",
  form_url = ""
)
}
\arguments{
\item{team_email}{Provide your team e-mail address as a string}

\item{form_url}{}
}
\description{
Create the standard DfE R-Shiny support and feedback dashboard panel.
}
