% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_panel.R
\name{support_panel}
\alias{support_panel}
\title{Support panel}
\usage{
support_panel(
  team_email = "",
  repo_name = "",
  ees_publication = TRUE,
  publication_name = NULL,
  publication_slug = "",
  alt_href = NULL,
  form_url = NULL,
  cookie_status_output = "cookie_status"
)
}
\arguments{
\item{team_email}{Your team e-mail address, must be a education.gov.uk email}

\item{repo_name}{The repository name as listed on GitHub}

\item{ees_publication}{Whether the parent publication is hosted on Explore
Education Statistics}

\item{publication_name}{The parent publication name}

\item{publication_slug}{The parent publication slug on Explore Education
Statistics}

\item{alt_href}{Alternative link to the parent publication (if not hosted on
Explore Education Statistics)}

\item{form_url}{URL to a feedback form for the dashboard}

\item{cookie_status_output}{Name of cookie status output object, often
"cookie_status"}
}
\value{
a standardised panel for a public R Shiny dashboard in DfE
}
\description{
Create the standard DfE R-Shiny support and feedback dashboard panel.
}
\examples{
support_panel(
  team_email = "my.team@education.gov.uk",
  repo_name = "https://github.com/dfe-analytical-services/my-repo",
  publication_name = "My publication title",
  publication_slug = "my-publication-title",
  form_url = "www.myform.com",
  cookie_status_output = "cookie_status"
)
}
