% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_panel.R
\name{support_panel}
\alias{support_panel}
\title{Support panel}
\usage{
support_panel(
  team_email = "",
  repo_name = "",
  ees_publication = TRUE,
  publication_name = NULL,
  publication_slug = "",
  alt_href = NULL,
  form_url = NULL
)
}
\arguments{
\item{team_email}{Your team e-mail address, must be a education.gov.uk email}

\item{repo_name}{The repository URL, must be a valid URL for the
dfe-analytical-services GitHub area}

\item{ees_publication}{Whether the parent publication is hosted on Explore
Education Statistics}

\item{publication_name}{The parent publication name}

\item{publication_slug}{The parent publication slug on Explore Education
Statistics}

\item{alt_href}{Alternative link to the parent publication (if not hosted on
Explore Education Statistics)}

\item{form_url}{URL for a feedback form for the dashboard}
}
\value{
a HTML div, containing standard support content for a public R Shiny
dashboard in DfE
}
\description{
Create the standard DfE R-Shiny support and feedback dashboard panel.
}
\examples{
support_panel(
  team_email = "my.team@education.gov.uk",
  repo_name = "https://github.com/dfe-analytical-services/my-repo",
  publication_name = "My publication title",
  publication_slug = "my-publication-title",
  form_url = "www.myform.com"
)

# Often you will use this inside a set of navigation tabs, e.g.
shiny::navlistPanel(
  "",
  id = "navlistPanel",
  widths = c(2, 8),
  well = FALSE,
  ## Support panel --------------------------------------------------------
  shiny::tabPanel(
    value = "support_panel",
    "Support and feedback",
    support_panel(
      team_email = "explore.statistics@education.gov.uk",
      repo_name = "https://github.com/dfe-analytical-services/dfeshiny/",
      form_url = "https://forms.office.com"
    )
  )
)
}
