% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_panel.R
\name{support_panel}
\alias{support_panel}
\title{Support panel}
\usage{
support_panel(
  team_email = "",
  repo_name = "",
  ees_publication = TRUE,
  publication_name = NULL,
  publication_slug = "",
  alt_href = NULL,
  form_url = NULL,
  custom_data_info = NULL,
  extra_text = NULL
)
}
\arguments{
\item{team_email}{Your team e-mail address, must be a education.gov.uk email}

\item{repo_name}{The repository URL, must be a valid URL for the
dfe-analytical-services GitHub area}

\item{ees_publication}{Whether the parent publication is hosted on Explore
Education Statistics}

\item{publication_name}{The parent publication name}

\item{publication_slug}{The parent publication slug on Explore Education
Statistics}

\item{alt_href}{Alternative link to the parent publication (if not hosted on
Explore Education Statistics)}

\item{form_url}{URL for a feedback form for the dashboard}

\item{custom_data_info}{A single vector or a combined vector wrapped in
\code{shiny::tagList()} instead of c() for custom text to go under the
"Find out more information on the data" heading.}

\item{extra_text}{Add extra paragraphs to the page before the "Contact us" section.
Use \code{dfeshiny::section_tags()} to specify the heading and body.
Look at examples to see how to add one or multiple sections.}
}
\value{
a HTML div, containing standard support content for a public R Shiny
dashboard in DfE
}
\description{
Create the standard DfE R-Shiny support and feedback dashboard panel.
}
\examples{
support_panel(
  team_email = "my.team@education.gov.uk",
  repo_name = "https://github.com/dfe-analytical-services/my-repo",
  publication_name = "My publication title",
  publication_slug = "my-publication-title",
  form_url = "www.myform.com"
)

# Often you will use this inside a set of navigation tabs, e.g.
shiny::navlistPanel(
  "",
  id = "navlistPanel",
  widths = c(2, 8),
  well = FALSE,
  ## Support panel --------------------------------------------------------
  shiny::tabPanel(
    value = "support_panel",
    "Support and feedback",
    support_panel(
      team_email = "explore.statistics@education.gov.uk",
      repo_name = "https://github.com/dfe-analytical-services/dfeshiny/",
      form_url = "https://forms.office.com"
    )
  )
)

# Example for adding custom text

# Adding custom text to the feedback section only

support_panel(
  team_email = "my.team@education.gov.uk",
  repo_name = "https://github.com/dfe-analytical-services/my-repo",
  publication_name = "My publication title",
  publication_slug = "my-publication-title",
  form_url = "www.myform.com",
  custom_data_info = "This is a sentence to test the ability to add custom text in the tab",
)

# Adding custom text that includes mixed elements to feedback section

support_panel(
  team_email = "my.team@education.gov.uk",
  repo_name = "https://github.com/dfe-analytical-services/my-repo",
  publication_name = "My publication title",
  publication_slug = "my-publication-title",
  form_url = "www.myform.com",
  custom_data_info = shiny::tagList(
    "Please email results to",
    external_link(
      href = paste0("mailto:", "team@education.gov.uk"),
      link_text = "team@education.gov.uk",
      add_warning = FALSE
    )
  )
)
# Example for adding custom sections

# Adding one section
support_panel(
  team_email = "my.team@education.gov.uk",
  repo_name = "https://github.com/dfe-analytical-services/my-repo",
  custom_data_info = "This is a sentence to test the ability to add custom text in the tab",
  extra_text = section_tags(
    heading = "heading",
    body = "this is a body"
  )
)
# Adding two sections
support_panel(
  team_email = "my.team@education.gov.uk",
  repo_name = "https://github.com/dfe-analytical-services/my-repo",
  extra_text = c(
    section_tags(
      heading = "heading",
      body = "this is a body"
    ),
    section_tags(
      heading = "heading 2",
      body = "this is another example of a text"
    )
  )
)

# Adding a section with a `shiny::tagList()` in
# the `dfeshiny::section_tags()`

support_panel(
  team_email = "my.team@education.gov.uk",
  repo_name = "https://github.com/dfe-analytical-services/my-repo",
  extra_text = c(
    section_tags(
      heading = "Heading",
      body = shiny::tagList(
        "Please email results to",
        external_link(
          href = paste0("mailto:", "team@education.gov.uk"),
          link_text = "team@education.gov.uk",
          add_warning = FALSE
        )
      )
    )
  )
)
}
\seealso{
\code{\link[=section_tags]{section_tags()}}
}
