% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_panels.R
\name{support_panel}
\alias{support_panel}
\title{Support panel}
\usage{
support_panel(
  team_email = "",
  repo_name = "",
  ees_publication = TRUE,
  publication_name = NULL,
  publication_stub = "",
  alt_href = NULL,
  form_url = NULL
)
}
\arguments{
\item{team_email}{Your team e-mail address as a string}

\item{repo_name}{The repository name as listed on GitHub}

\item{ees_publication}{Whether the parent publication is hosted on Explore Education Statistics}

\item{publication_name}{The parent publication name}

\item{publication_stub}{The parent publication stub on Explore Education Statistics}

\item{alt_href}{Alternative link to the parent publication (if not hosted on Explore Education Statistics)}

\item{form_url}{URL to a feedback form for the dashboard}
}
\description{
Create the standard DfE R-Shiny support and feedback dashboard panel.
}
