% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_code.R
\name{tidy_code}
\alias{tidy_code}
\title{Tidy code}
\usage{
tidy_code(subdirs = c("R", "tests"))
}
\arguments{
\item{subdirs}{List of sub-directories to
(recursively search for R scripts to be styled)}
}
\value{
True or False value based on if scripts were changed
}
\description{
Script to apply styler code styling to scripts within the
shiny directory structure.
}
\examples{
tidy_code()
}
